move pydantic base models to separate file
